string: arm: Fix CFI auto-alignment issues.
The use of the PAC_CFI_ADJ macro for calculating the effect of pushing
the IP register onto the stack assumes that pushing the register is
always optional and is always supressed when PAC_LEAF_PUSH_IP is set
to 0. This leads to CFI alignment issues for functions where the IP
register is clobbered and thus where IP is always pushed to the stack
in the function prologue.

This patch introduces a new macro PAC_CFI_ADJ_DEFAULT whose value is
never zeroed when PAC signing is requested, irrespective of the
PAC_LEAF_PUSH_IP settings.

Example:
* HAVE_PAC_LEAF == 1 && PAC_LEAF_PUSH_IP == 1:
    PAC_CFI_ADJ = 4
    PAC_CFI_ADJ_DEFAULT = 4

* HAVE_PAC_LEAF == 1 && PAC_LEAF_PUSH_IP == 0:
    PAC_CFI_ADJ = 0
    PAC_CFI_ADJ_DEFAULT = 4

Built w/ arm-none-linux-gnueabihf, ran make check-string w/ qemu-arm-static.